+++ test/algorithms/indexing/tst_assign_indices.py
@@ -77,7 +77,7 @@ def run(space_group_info):
   miller_set = miller.set(
     crystal_symmetry, miller_indices, anomalous_flag=True)
   predicted_reflections['xyzobs.mm.value'] = predicted_reflections['xyzcal.mm']
-  predicted_reflections['id'] = flex.size_t(len(predicted_reflections), 0)
+  predicted_reflections['id'] = flex.int(len(predicted_reflections), 0)
   from dials.algorithms.indexing.indexer import indexer_base
   indexer_base.map_centroids_to_reciprocal_space(
     predicted_reflections, sweep.get_detector(), sweep.get_beam(),
